+++ src/stdlib/text.c
@@ -1124,7 +1124,7 @@ OptionalInt_t Text$find(Text_t text, Text_t target, Int_t start) {
     if (text.length < target.length) return NONE_INT;
     if (target.length <= 0) return I(1);
     TextIter_t text_state = NEW_TEXT_ITER_STATE(text), target_state = NEW_TEXT_ITER_STATE(target);
-    for (int64_t i = Int64$from_int(start, false) - 1; i < text.length - target.length + 1; i++) {
+    for (int64_t i = Int64$from_int(start, false) - 1; i < (int64_t)text.length - target.length + 1; i++) {
         if (_matches(&text_state, &target_state, i)) {
             return Int$from_int64(i + 1);
         }
@@ -1444,16 +1444,16 @@ double Text$distance(Text_t a, Text_t b, Text_t language) {
     // for letters that are the same, but with different casing.
     double *distances = GC_MALLOC_ATOMIC(sizeof(uint32_t) * (a.length + 1) * (b.length + 1));
 #define DIST(x, y) distances[(x) * b.length + (y)]
-    for (int64_t i = 0; i <= a.length; i++)
+    for (int64_t i = 0; i <= (int64_t)a.length; i++)
         DIST(i, 0) = i;
-    for (int64_t j = 0; j <= b.length; j++)
+    for (int64_t j = 0; j <= (int64_t)b.length; j++)
         DIST(0, j) = j;
 
     TextIter_t a_state = NEW_TEXT_ITER_STATE(a);
     TextIter_t b_state = NEW_TEXT_ITER_STATE(b);
     const char *uc_language = Text$as_c_string(language);
-    for (int64_t i = 1; i <= a.length; i++) {
-        for (int64_t j = 1; j <= b.length; j++) {
+    for (int64_t i = 1; i <= (int64_t)a.length; i++) {
+        for (int64_t j = 1; j <= (int64_t)b.length; j++) {
             int32_t ai = Text$get_grapheme_fast(&a_state, i - 1);
             int32_t bj = Text$get_grapheme_fast(&b_state, j - 1);
             double cost = (double)(ai != bj);
